// src/interfaces/Colors.ts
export interface Colors {
  background: string;
  point: string;
  primary: string;
  secondary: string;
  lace: LaceColor;
  metal: MetalColor;
  sole: SoleColor;
}

export enum LaceColor {
  WHITE = "#EEEEEE",
  BLACK = "#111111",
}

export enum MetalColor {
  WHITE = "#EEEEEE",
  GRAY = "#999999",
  BLACK = "#111111",
}

export enum SoleColor {
  WHITE = "#EEEEEE",
  BLACK = "#111111",
}